Feat/show rssi battery and status (#46)
* feat(protocol): add error flag constants for node health conditions

Define ERR_FLAG_* constants in message.h for heartbeat error reporting:
- SENSOR_FAILURE: temp/humidity sensor not responding
- FLASH_FAILURE: external flash not responding
- FLASH_FULL: flash storage >90% full
- PMU_FAILURE: PMU communication failure
- BATTERY_LOW: battery <20%
- BATTERY_CRITICAL: battery <10%
- RTC_NOT_SYNCED: RTC never synchronized
- RADIO_ISSUE: LoRa transmission issues

* feat(firmware): implement error detection in sensor mode

Add collectErrorFlags() and getBatteryLevel() methods to sensor mode:
- Track sensor read validity for error reporting
- Check flash status and capacity (>90% triggers warning)
- Check PMU availability
- Check RTC sync status
- Check battery level for low/critical warnings
- Update sendHeartbeat() to use actual error flags and RSSI

* feat(firmware): forward heartbeat status to API via UART

Update handleHeartbeat() to forward node status to Raspberry Pi:
- Format: HEARTBEAT <addr> <device_id> <battery> <error_flags> <signal> <uptime>
- Include actual RSSI from received message
- Look up device_id from address manager

* feat(api): add node status storage and endpoints

- Add node_status table to database for storing heartbeat data
- Add update_node_status(), get_node_status(), get_all_node_status() methods
- Parse HEARTBEAT messages from hub in serial interface
- Update GET /api/nodes to include status field
- Add GET /api/nodes/<addr>/status endpoint

* feat(dashboard): add node status indicators with Lucide icons

- Add lucide-react for battery, signal, and health icons
- Add NodeStatus type and error flag constants to types
- Add helper functions: parseErrorFlags, getHealthStatus, getSignalQuality, etc.
- Create BatteryGauge component with level-based icons
- Create SignalStrength component with RSSI-based signal bars
- Create HealthStatus component with tooltip for error details
- Add status row to NodeCard (battery, signal, health)
- Add status panel to NodeDetail with uptime and error badges

* Formatting

* fix(api): rename addr parameter to address

Per CLAUDE.md design principle: "Use full words not abbreviations
(e.g address vs addr)". Renamed all occurrences of 'addr' to
'address' in route parameters and function bodies. Protocol format
comments describing serial interface remain unchanged.

* fix(firmware): correct RSSI calculation in heartbeat

RSSI values are negative dBm (e.g., -70 dBm), so the condition
"getRssi() > 0" was always false, causing the hardcoded fallback
value of 70 to always be used.

Now properly checks for valid negative RSSI range (-120 to 0 dBm)
and converts to absolute value for the HeartbeatPayload which
stores signal_strength as uint8_t.

* feat(firmware): extend error_flags and add TX/timeout detection

Extend error_flags from uint8_t to uint16_t to support additional
status indicators:

- ERR_FLAG_TX_FAILURES (0x80): Set after 3+ consecutive batch
  transmission failures. Indicates sensor has data but cannot
  deliver it to the hub.

- ERR_FLAG_HIGH_TIMEOUTS (0x100): Set when >25% of reliable/critical
  messages are timing out (requires minimum 10 message sample size).
  Uses NetworkStats to track timeout rates.

Renamed ERR_FLAG_RADIO_ISSUE to ERR_FLAG_TX_FAILURES for clarity.

Implementation adds consecutive_tx_failures_ counter that increments
on batch ACK failure and resets on success.

* style: apply clang-format

* fix(firmware): use sensor's RSSI from payload instead of hub measurement

The hub was calling lora_.getRssi() after receiving a heartbeat, which
reads the RSSI of the last received packet from a hardware register.
This has potential race conditions with multiple nodes.

Now uses payload->signal_strength from the heartbeat, which the sensor
captured when it received a message from the hub. This is more reliable
and represents the sensor's view of link quality.

The value is converted from absolute (stored in payload) back to
negative dBm for the API.

* feat(firmware): add pending_records to heartbeat payload

Extend HeartbeatPayload struct with pending_records field to report
the count of untransmitted sensor records in flash backlog.

- Add pending_records (uint16_t) to HeartbeatPayload in message.h
- Update sendHeartbeat() signature with optional pending_records param
- Populate pending_records from flash_buffer in sensor_mode.cpp
- Forward pending_records in hub serial output to RasPi

* feat(api): add pending_records and hub_queue_count to node status

Store and expose backlog information for dashboard display:

- Add pending_records column to node_status table schema
- Update update_node_status(), get_node_status(), get_all_node_status()
- Parse optional pending_records field from HEARTBEAT serial line
- Add hub_queue_count via GET_QUEUE command (opt-in with include_queue param)

* feat(dashboard): add BacklogStatus component to NodeCard

Display sensor backlog and hub queue counts in node cards:

- Add pending_records to NodeStatus interface
- Add hub_queue_count to Node interface
- Add getBacklogStatus() helper for color coding
- Create BacklogStatus component with Database/CloudUpload icons
- Add BacklogStatus to NodeCard status row

* chore: add formatting tools and configuration

- Add ruff to Python dev dependencies with formatting config
- Add prettier to dashboard dev dependencies
- Create .prettierrc with consistent style settings

Note: pre-commit hook installed locally in .git/hooks/pre-commit

* chore: add sharable pre-commit hook for auto-formatting

Move pre-commit hook to scripts/ for version control. New developers
can enable it by running:

  ln -sf ../../scripts/pre-commit .git/hooks/pre-commit

* style: update hub_mode.cpp formatting

* docs(claude): update build/flash instructions for pico-sdk paths

- Update CMake path to use bundled pico-sdk version
- Add hardware variant flag to build commands
- Add OpenOCD flash command example
- Document /bramble-build and /bramble-flash skills for RP2040
- Clarify STM32 PMU requires manual build/flash


* fix(hub): measure RSSI at hub instead of using stale node payload

The hub now captures RSSI when receiving heartbeat packets rather than
using the stale/incorrect value from the node's payload. This provides
accurate node→hub link quality measurement.

Previously the node reported RSSI from previous receive cycles (or 0),
which measured the wrong direction (hub→node) and was delayed.


* feat(sensor): add lazy init with power-on delay and fix error timing

- Defer sensor initialization until first read attempt
- Add 1s power-on stabilization delay before I2C init
- Add 20s delay on first boot (PMU time invalid) for programming
- Fix: only report ERR_FLAG_SENSOR_FAILURE after attempting a read

The lazy init addresses I2C sensor issues when power cycling rapidly.
The error flag fix prevents false sensor failure reports in heartbeats
sent before the first sensor read attempt.


* fix(sensor): calculate uptime from boot time instead of parameter

sendHeartbeat() was often called with 0 as current_time, resulting
in uptime always being reported as 0. Now uses actual boot time
via to_ms_since_boot() regardless of parameter value.


* feat(sensor): persist initial boot timestamp for uptime across sleep cycles

Store initial_boot_timestamp in flash metadata to track true uptime since
power-on (battery connection) rather than since last wake from sleep.

- Add initial_boot_timestamp field to SensorFlashMetadata
- Add getInitialBootTimestamp/setInitialBootTimestamp to SensorFlashBuffer
- Set timestamp when RTC first becomes valid after cold start
- Calculate uptime as (current_rtc_time - initial_boot_timestamp)
- Fix race condition: set rtc_synced_ before sendHeartbeat() to avoid
  false ERR_FLAG_RTC_NOT_SYNCED when RTC is valid from PMU


---------

+++ src/lora/message.cpp
@@ -67,7 +67,7 @@ size_t MessageHandler::createActuatorMessage(uint16_t src_addr, uint16_t dst_add
 
 size_t MessageHandler::createHeartbeatMessage(uint16_t src_addr, uint16_t dst_addr, uint8_t seq_num,
                                               uint8_t battery_level, uint8_t signal_quality,
-                                              uint32_t uptime_seconds, uint8_t status_flags,
+                                              uint32_t uptime_seconds, uint16_t status_flags,
                                               uint8_t *buffer)
 {
     HeartbeatPayload payload = {.uptime_seconds = uptime_seconds,
